refactor bucket() & related tests into core
