+++ cli/src/cli/subcommand/range/bybucket.rs
@@ -38,7 +38,7 @@ where
     let mut buckets: Vec<Option<i64>> = interval.iter().map(|_| None).collect();
 
     for item in input.into_iter() {
-        bucket(interval, item as f64).and_then(|bucket| {
+        Interval::bucket(interval, item as f64).and_then(|bucket| {
             buckets
                 .get_mut(bucket)
                 .filter(|current| select(item, **current))
@@ -47,55 +47,4 @@ where
     }
 
     buckets.into_iter().filter_map(|opt| opt).collect()
-}
-
-fn bucket(interval: &Interval, number: f64) -> Option<usize> {
-    if number < interval.low() || number >= interval.high() {
-        return None;
-    }
-
-    let bucket = f64::ln(number - interval.low() + 1.0)
-        / f64::ln(interval.high() - interval.low() + 1.0)
-        * interval.count() as f64;
-
-    Some(bucket.trunc() as usize)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_bucket() {
-        let data = vec![
-            TestData::new(Interval::new(1.0, 10.0, 5).unwrap(), 8.0, Some(4)),
-            TestData::new(Interval::new(30.0, 100.0, 10).unwrap(), 1000.0, None),
-            TestData::new(Interval::new(30.0, 100.0, 10).unwrap(), 0.0, None),
-            TestData::new(Interval::new(30.0, 100.0, 10).unwrap(), 10.0 * 10.0, None),
-            TestData::new(Interval::new(-100.0, 100.0, 10).unwrap(), 0.0, Some(8)),
-            TestData::new(Interval::new(-100.0, 100.0, 10).unwrap(), -100.0, Some(0)),
-        ];
-
-        for test in data {
-            let actual = bucket(&test.interval, test.number);
-
-            assert_eq!(test.expected, actual)
-        }
-    }
-
-    struct TestData {
-        interval: Interval,
-        number: f64,
-        expected: Option<usize>,
-    }
-
-    impl TestData {
-        fn new(interval: Interval, number: f64, expected: Option<usize>) -> Self {
-            Self {
-                interval,
-                number,
-                expected,
-            }
-        }
-    }
 }
